% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChooseK.R
\name{ChooseK}
\alias{ChooseK}
\alias{plot.ChooseK}
\title{Choose the Number of Clusters K for (Sparse) Weighted K-Means Clustering}
\usage{
ChooseK(x, nClusters = 2:6, nperms = 20, nstart = 20)

\method{plot}{ChooseK}(x, ...)
}
\arguments{
\item{x}{An \emph{n} by \emph{p} numeric data matrix, and \emph{n} is the number of observations and \emph{p} the number of features.}

\item{nClusters}{a candidate sequence of K. Default is \code{2:6}.}

\item{nperms}{Number of permutations. Default is \code{20}.}

\item{nstart}{The number of initial random sets chosen from (distinct) rows in \code{x}. Omitted if \code{centers} is provided. Default is 20.}

\item{...}{unused.}
}
\value{
The function returns a list of the following components:
\item{OptimalK}{The optimal number of clusters chosen from \code{nClusters}.}
\item{plotinfo}{A list containing the information needed in S3 method \code{plot}.}
}
\description{
The number of clusters K should be determined before the clustering method is performed. 
A permutation approach using Gap Statistic is used.
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: plot the Gap statistic of each candicate number of clusters K.
}}

\references{
Robert, T. \emph{et al.} (2001). Estimating the number of clusters in a data set via the gap statistic. \emph{Journal of the Royal Statistical Society: Series B (Statistical Methodology)}, \bold{63(2)}, 411-423.
}
\seealso{
Other sparse weighted K-Means functions: \code{\link{KMeansSparseCluster.permute.weight}},
  \code{\link{KMeansSparseCluster.weight}},
  \code{\link{kmeans.weight.tune}},
  \code{\link{kmeans.weight}}
}
\author{
Wenyu Zhang
}
\concept{sparse weighted K-Means functions}
\keyword{Clustering}
\keyword{K-Means}
\keyword{Parameter}
\keyword{Tuning}
\keyword{Weighted}
