% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMeansSparseCluster_permute_weight.R
\name{KMeansSparseCluster.permute.weight}
\alias{KMeansSparseCluster.permute.weight}
\title{Choose Sparsity Parameter for Sparse Weighted K-Means Clustering}
\usage{
KMeansSparseCluster.permute.weight(x, K = NULL, weight = NULL,
  nperms = 20, nstart = 20, wbounds = NULL, silent = TRUE,
  nvals = 10, centers = NULL)
}
\arguments{
\item{x}{An \emph{n} by \emph{p} numeric data matrix, and \emph{n} is the number of observations and \emph{p} the number of features.}

\item{K}{The number of clusters. Omitted if \code{centers} are provided.}

\item{weight}{A vector of \emph{n} positive elements representing weights on observations.}

\item{nperms}{Number of permutations. Default is 20.}

\item{nstart}{The number of initial random sets chosen from (distinct) rows in \code{x}. Omitted if \code{centers} is provided. Default is 20.}

\item{wbounds}{A single L1 bound on w (the feature weights), or a vector of L1 bounds on w. If wbound is small,
    then few features will have non-zero weights. If wbound is large then all
  features will have non-zero weights. Should be greater than 1.}

\item{silent}{Print out progress?}

\item{nvals}{The number of candidate tuning parameter values. Omitted if \code{wbounds} is given.}

\item{centers}{A \emph{K} by \emph{p} matrix indicating initial (distinct) cluster centers.}
}
\value{
\item{gaps}{The gap statistics obtained (one for each of the tuning
    parameters tried). If O(s) is the objective function evaluated at
    the tuning parameter s, and O*(s) is the same quantity but for the
    permuted data, then Gap(s)=log(O(s))-mean(log(O*(s))).}
  \item{sdgaps}{The standard deviation of log(O*(s)), for each value of the
    tuning parameter s.}
  \item{nnonzerows}{The number of features with non-zero weights, for
    each value of the tuning parameter.}
  \item{wbounds}{The tuning parameters considered.}
  \item{bestw}{The value of the tuning parameter corresponding to the
    highest gap statistic.}
}
\description{
The sparsity parameter controls the L1 bound on w, the feature weights.
 A permutation approach is used to select the sparsity parameter.
}
\examples{
# generate data
set.seed(1)
require(mvtnorm)
n <- 60  #sample size
p <- 1000 #dimension of features
q <- 50  #dimension of cluster-specific features
mu <- 0.8
MU <- c(0,-mu,mu)
sigma0 <- 5
data <- rbind(rmvnorm(n/3,rep(0,p)),rmvnorm(n/3,c(rep(-mu,q),rep(0,p-q))),
rmvnorm(n/3,c(rep(mu,q),rep(0,p-q))))
# add noise to 10 random observations
noisy.lab <- sample(n,10)
for (k in 1:3){
check <- (noisy.lab<n*k/3+1) & (noisy.lab>n/3*(k-1))
temp.lab <- noisy.lab[check]
num <- length(temp.lab)
if(any(check))
  data[temp.lab,] <- rmvnorm(num,c(rep(MU[k],q),rep(0,p-q)),sigma = diag(sigma0,p))
}
# run kmeans.weight.tune to tune weight parameter U
res.tuneU <- kmeans.weight.tune(data,K=3,noisy.lab=noisy.lab)
plot(res.tuneU)
weight <- res.tuneU$bestweight
# run KMeansSparseCluster.weight.permute to tune sparsity parameter s
res.tunes <- KMeansSparseCluster.permute.weight(data,K=3,weight=weight)
res <- KMeansSparseCluster.weight(data,K=3,weight=weight,wbounds=res.tunes$bestw)
}
\seealso{
Other sparse weighted K-Means functions: \code{\link{KMeansSparseCluster.weight}},
  \code{\link{kmeans.weight.tune}},
  \code{\link{kmeans.weight}}
}
\author{
Wenyu Zhang, Jiaxuan Wangwu and Zhixiang Lin
}
\concept{sparse weighted K-Means functions}
\keyword{Clustering}
\keyword{K-Means}
\keyword{Parameter}
\keyword{Sparse}
\keyword{Tuning}
\keyword{Weighted}
