% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NormalDisData}
\alias{NormalDisData}
\title{A Dataset Containing 3 Clusters of Normal Distribution Data}
\format{A list with three items:
\describe{
  \item{data}{a 60 by 500 matrix, indicating 60 observations with 500 features. The first 
  50 features are cluster-specific.}
  \item{nCluster}{true number of clusters.}
  \item{true.label}{a 60-dimension integer vector indicating the cluster each observation 
  should be assigned to.}
  \item{noisy.label}{a 10-dimension integer vector indicating the positions of noisy 
  observations. The noisy observations share the same mean with normal ones but have larger 
  variance.}
}}
\source{
This dataset is generated by the code shown in the example.
}
\usage{
NormalDisData
}
\description{
A simulated dataset used in the `SWKM` vignette for illustration to perform sparse weighted
K-Means clustering.
}
\examples{
\dontrun{
# this is the code to generate this dataset.
set.seed(1)
require(mvtnorm)
n <- 60  #sample size
p <- 500 #dimension of features
q <- 50  #dimension of cluster-specific features
mu <- 0.8
MU <- c(0,-mu,mu)
sigma0 <- 5
data <- rbind(rmvnorm(n/3,rep(0,p)),rmvnorm(n/3,c(rep(-mu,q),rep(0,p-q))),
              rmvnorm(n/3,c(rep(mu,q),rep(0,p-q))))
# add noise to 10 random observations
noisy.lab <- sample(n,10)
for (k in 1:3){
  check <- (noisy.lab<n*k/3+1) & (noisy.lab>n/3*(k-1))
  temp.lab <- noisy.lab[check]
  num <- length(temp.lab)
  if(any(check))
    data[temp.lab,] <- rmvnorm(num,c(rep(MU[k],q),rep(0,p-q)),sigma = diag(sigma0,p))
}
}
}
\keyword{datasets}
