% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_weight.R
\name{kmeans.weight}
\alias{kmeans.weight}
\title{Weighted K-Means Clustering with Weights on Observations}
\usage{
kmeans.weight(x, K = NULL, weight = NULL, centers = NULL,
  nstart = 20, algorithm = "Hartigan-Wong")
}
\arguments{
\item{x}{An \emph{n} by \emph{p} numeric data matrix, and \emph{n} is the number of observations and \emph{p} the number of features.}

\item{K}{The number of clusters. Omitted if \code{centers} are provided.}

\item{weight}{A vector of \emph{n} positive elements representing weights on observations.}

\item{centers}{A \emph{K} by \emph{p} matrix indicating initial (distinct) cluster centers.}

\item{nstart}{The number of initial random sets chosen from (distinct) rows in \code{x}. Omitted if \code{centers} is provided. Default is 20.}

\item{algorithm}{Character; either "\code{Hartigon-Wong}" or "\code{Forgy}". Default is "\code{Hartigon-Wong}".}
}
\value{
The function returns a list of the following components:
\item{centers}{the centers of the clustering result.}
\item{cluster}{a vector of integers (from \code{1:k}) indicating the cluster to which each observation is allocated.}
\item{weight}{a vector of non-zero weights in the input vector \code{weight}.}
\item{wcss}{normalized within-cluster sum of squares, i.e. the objective divided by \code{sum(weight)}.}
}
\description{
Perform K-Means algorithm on observations with given weights.
}
\examples{
# generate data
set.seed(1)
require(mvtnorm)
n <- 60  #sample size
p <- 1000 #dimension of features
q <- 50  #dimension of cluster-specific features
mu <- 0.8
MU <- c(0,-mu,mu)
sigma0 <- 5
data <- rbind(rmvnorm(n/3,rep(0,p)),rmvnorm(n/3,c(rep(-mu,q),rep(0,p-q))),
rmvnorm(n/3,c(rep(mu,q),rep(0,p-q))))
# add noise to 10 random observations
noisy.lab <- sample(n,10)
for (k in 1:3){
check <- (noisy.lab<n*k/3+1) & (noisy.lab>n/3*(k-1))
temp.lab <- noisy.lab[check]
num <- length(temp.lab)
if(any(check))
  data[temp.lab,] <- rmvnorm(num,c(rep(MU[k],q),rep(0,p-q)),sigma = diag(sigma0,p))
}
# apply kmeans.weight.tune to tune weight parameter U
res.tuneU <- kmeans.weight.tune(data,K=3,noisy.lab=noisy.lab)
plot(res.tuneU)
# perform weighted K-Means to get the clustering result
res <- kmeans.weight(data,K=3,weight=res.tuneU$bestweight)
}
\seealso{
Other sparse weighted K-Means functions: \code{\link{KMeansSparseCluster.permute.weight}},
  \code{\link{KMeansSparseCluster.weight}},
  \code{\link{kmeans.weight.tune}}
}
\author{
Wenyu Zhang
}
\concept{sparse weighted K-Means functions}
\keyword{Clustering}
\keyword{K-Means}
\keyword{Weighted}
