% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_weight_tune.R
\name{kmeans.weight.tune}
\alias{kmeans.weight.tune}
\alias{plot.kmeans.weight.tune}
\title{Choose Tuning Parameter for Weighted K-Means Clustering}
\usage{
kmeans.weight.tune(x, K = NULL, noisy.lab = NULL, weight.seq = NULL,
  nperms = 20, centers = NULL, nstart = 20,
  algorithm = "Hartigan-Wong")

\method{plot}{kmeans.weight.tune}(x, ...)
}
\arguments{
\item{x}{An \emph{n} by \emph{p} numeric data matrix, and \emph{n} is the number of observations and \emph{p} the number of features.}

\item{K}{The number of clusters. Omitted if \code{centers} are provided.}

\item{noisy.lab}{A vector indicating the row positions of noisy observations. Omitted if \code{weight.seq} is provided.}

\item{weight.seq}{A candidate weight matrix, each row indicating one candidate weight vector. If \code{NULL},
the function will assign a sequence of candidate weight \code{c(1,0.8,0.5,0.2,0.1,0.08,0.05,0.02,0.01,0.005,0.001)}
to noisy observations by default.}

\item{nperms}{Number of permutations. Default is 20.}

\item{centers}{A \emph{K} by \emph{p} matrix indicating initial (distinct) cluster centers.}

\item{nstart}{The number of initial random sets chosen from (distinct) rows in \code{x}. Omitted if \code{centers} is provided. Default is 20.}

\item{algorithm}{Character; either "\code{Hartigan-Wong}" or "\code{Forgy}". Default is "\code{Hartigon-Wong}".}

\item{...}{unused.}
}
\value{
The function returns a list of the following components:
\item{gaps}{The gap statistics obtained (one for each of the candicate weights tried).
If \eqn{O(U)} is the objective function evaluated at the tuning parameter \eqn{U},
and \eqn{O*(U)} is the same quantity but for the permuted weights,
then Gap(\eqn{U})=mean(log\eqn{(O*(U))})-log\eqn{(O(U))}.}
\item{sdgaps}{The standard deviation of log\eqn{(O*(U))}}
\item{bestweight}{The best weight chosen by this method among all the candidate weights.}
}
\description{
The tuning parameter controls the weight of noisy observations.
A permutation approach is used to select the tuning parameter.
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: plot the Gap statistic of each candicate weight vector.
}}

\examples{
# generate data
set.seed(1)
require(mvtnorm)
n <- 60  #sample size
p <- 1000 #dimension of features
q <- 50  #dimension of cluster-specific features
mu <- 0.8
MU <- c(0,-mu,mu)
sigma0 <- 5
data <- rbind(rmvnorm(n/3,rep(0,p)),rmvnorm(n/3,c(rep(-mu,q),rep(0,p-q))),
rmvnorm(n/3,c(rep(mu,q),rep(0,p-q))))
# add noise to 10 random observations
noisy.lab <- sample(n,10)
for (k in 1:3){
check <- (noisy.lab<n*k/3+1) & (noisy.lab>n/3*(k-1))
temp.lab <- noisy.lab[check]
num <- length(temp.lab)
if(any(check))
  data[temp.lab,] <- rmvnorm(num,c(rep(MU[k],q),rep(0,p-q)),sigma = diag(sigma0,p))
}
# apply kmeans.weight.tune to tune weight parameter U
res.tuneU <- kmeans.weight.tune(data,K=3,noisy.lab=noisy.lab)
plot(res.tuneU)
# perform weighted K-Means to get the clustering result
res <- kmeans.weight(data,K=3,weight=res.tuneU$bestweight)
}
\seealso{
Other sparse weighted K-Means functions: \code{\link{KMeansSparseCluster.permute.weight}},
  \code{\link{KMeansSparseCluster.weight}},
  \code{\link{kmeans.weight}}
}
\author{
Wenyu Zhang
}
\concept{sparse weighted K-Means functions}
\keyword{Clustering}
\keyword{K-Means}
\keyword{Parameter}
\keyword{Tuning}
\keyword{Weighted}
